+++ src/Controller/AdminController.php
@@ -24,69 +24,9 @@ final class AdminController extends AbstractController
         return $this->render('admin/users.html.twig');
     }
 
-    #[Route('/admin/companies', name: 'admin_companies')]
-    public function companies(EntityManagerInterface $em): Response
-    {
-        $companies = $em->getRepository(Company::class)->findAll();
-
-        return $this->render('admin/companies.html.twig', [
-            'companies' => $companies
-        ]);
-    }
-
     #[Route('/admin/forum', name: 'admin_forum')]
     public function showForum(EntityManagerInterface $em): Response
     {
         return $this->render('admin/forum.html.twig');
-    }
-
-    #[Route('/admin/company/update', name: 'company_update', methods: ['POST'])]
-    public function updateCompany(Request $request, EntityManagerInterface $em): JsonResponse
-    {
-        try {
-            $data = json_decode($request->getContent(), true);
-
-            if (!$data || !isset($data['oldCompanyName']) || !isset($data['companyName'])) {
-                return new JsonResponse([
-                    'error' => 'Données invalides',
-                    'received' => $data
-                ], 400);
-            }
-
-            $company = $em->getRepository(Company::class)->findOneBy(['company_name' => $data['oldCompanyName']]);
-
-            if (!$company) {
-                return new JsonResponse([
-                    'error' => 'Entreprise introuvable',
-                    'searchedName' => $data['oldCompanyName']
-                ], 404);
-            }
-
-            // Conserver le logo existant si non fourni dans la mise à jour
-            $currentLogo = $company->getCompanyLogo();
-
-            $company->setCompanyName($data['companyName']);
-            $company->setCompanyDescription($data['companyDescription'] ?? '');
-            // Préserver le logo existant si pas de nouveau logo fourni
-            $company->setCompanyLogo($data['companyLogo'] ?? $currentLogo);
-
-            $em->flush();
-
-            return new JsonResponse([
-                'success' => true,
-                'data' => [
-                    'companyName' => $company->getCompanyName(),
-                    'companyDescription' => $company->getCompanyDescription(),
-                    'companyLogo' => $company->getCompanyLogo()
-                ]
-            ]);
-
-        } catch (\Exception $e) {
-            error_log('Erreur mise à jour entreprise: ' . $e->getMessage());
-            return new JsonResponse([
-                'error' => 'Erreur lors de la mise à jour',
-                'message' => $e->getMessage()
-            ], 500);
-        }
     }
 }
